ecommerce/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.0.0
stripe==11.3.0
python-dotenv==1.0.0
dj-database-url==2.2.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
